refactor : Redis 기반 좋아요 추가/삭제 원자성
- 좋아요 수 증가/감소 기능을 Redis INCR/DECR 기반으로 원자성 있게 처리
